feat: Support next-redux-wrapper v6 and v7
Since `getStaticProps()` and `getServerSideProps()` were introduced in Next.js v9.3, the API of
next-redux-wrapper has been adapted in next-redux-wrapper v6. As a reaction to this,
next-redux-cookie-wrapper has been rewritten to support the new API.

BREAKING CHANGE: This release is a full rewrite that drops Redux Persist and uses a custom
middleware instead. next-redux-cookie-wrapper is no longer a drop-in replacement for
next-redux-wrapper now but rather an extension to it, namely a Redux middleware to be used with
next-redux-wrapper. Hence, the API of v2 has fully changed since v1. Please follow the setup
instructions in the readme to set up v2 in your project.

Closes #13